Implements perimeter calculation for a rectangle
* implements calculate perimeter method inside the rectangle class
* changes signature to accept double and int values as well

+++ src/com/tw/bootcamp/problem1/App.java
@@ -5,6 +5,6 @@ import com.tw.bootcamp.problem1.shapes.Rectangle;
 public class App {
     static void main() {
         Rectangle rectangle = new Rectangle(4, 5);
-        int area = rectangle.calculateArea();
+        double area = rectangle.calculateArea();
     }
 }
